refactor: remove dataviewer component
