+++ package.json
@@ -8,7 +8,7 @@
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview",
-    "fix": "prettier . --write",
+    "format:fix": "prettier . --write",
     "prepare": "husky install"
   },
   "dependencies": {
@@ -34,5 +34,9 @@
     "prettier": "3.0.3",
     "typescript": "^5.0.2",
     "vite": "^4.4.8"
+  },
+  "lint-staged": {
+    "src/**/*.{js,jsx,ts,tsx}": "eslint --cache --fix",
+    "src/**/*.{js,jsx,ts,tsx,css,scss,md}": "prettier --write --ignore-unknown"
   }
 }
