+++ src/Components/DataVieewer/DataViewer.type.ts
@@ -1,0 +1,9 @@
+import type { PeopleResponse } from "../../apiTypes";
+type DataViewerProps = {
+  data: PeopleResponse[];
+  loadStatus: boolean;
+};
+
+type DataViewerState = {};
+
+export type { DataViewerProps, DataViewerState };
